+++ CHANGELOG.md
@@ -22,4 +22,5 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 ### Oneview Features supported
 - Enclosure    
 - FC network
-- Logical interconnect   
+- Logical interconnect
+- Scope